requirements.txt
Flask==3.1.0
python-dotenv==1.1.0
Requests==2.32.3
retrying==1.3.4
